iter-47: review fixes (BlockedIPs casing, non-negative parsing, double types)
- Rename BlockedIPs on the wire to preserve the IP acronym
- Clamp firewall numeric flags to non-negative via clap value_parser
- Type limit_rate_after / limit_rate_per_second as f64 to match spec
  (LimitRateAfter / LimitRatePerSecond are documented as double)

+++ crates/bunny-net-api/src/core/types.rs
@@ -978,14 +978,10 @@ pub struct PullZone {
     pub burst_size: Option<i32>,
     #[serde(default)]
     pub request_limit: Option<i32>,
-    /// The API returns this field as a float (e.g. `0.0`), so we store it as
-    /// `f64` to avoid deserialisation errors. Use `.round() as i32` when
-    /// passing to `UpdatePullZone::limit_rate_after`.
+    /// The API documents this field as `double`, so it is stored as `f64`.
     #[serde(default)]
     pub limit_rate_after: Option<f64>,
-    /// The API returns this field as a float (e.g. `0.0`), so we store it as
-    /// `f64` to avoid deserialisation errors. Use `.round() as i32` when
-    /// passing to `UpdatePullZone::limit_rate_per_second`.
+    /// The API documents this field as `double`, so it is stored as `f64`.
     #[serde(default)]
     pub limit_rate_per_second: Option<f64>,
     #[serde(rename = "ConnectionLimitPerIPCount", default)]
@@ -1402,7 +1398,7 @@ pub struct UpdatePullZone {
     pub blocked_countries: Option<Vec<String>>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub budget_redirected_countries: Option<Vec<String>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(rename = "BlockedIPs", skip_serializing_if = "Option::is_none")]
     pub blocked_ips: Option<Vec<String>>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub allowed_referrers: Option<Vec<String>>,
@@ -1434,9 +1430,9 @@ pub struct UpdatePullZone {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub request_limit: Option<i32>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub limit_rate_after: Option<i32>,
+    pub limit_rate_after: Option<f64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub limit_rate_per_second: Option<i32>,
+    pub limit_rate_per_second: Option<f64>,
     #[serde(
         rename = "ConnectionLimitPerIPCount",
         skip_serializing_if = "Option::is_none"
@@ -2143,13 +2139,13 @@ impl UpdatePullZone {
     }
 
     #[must_use]
-    pub fn limit_rate_after(mut self, v: i32) -> Self {
+    pub fn limit_rate_after(mut self, v: f64) -> Self {
         self.limit_rate_after = Some(v);
         self
     }
 
     #[must_use]
-    pub fn limit_rate_per_second(mut self, v: i32) -> Self {
+    pub fn limit_rate_per_second(mut self, v: f64) -> Self {
         self.limit_rate_per_second = Some(v);
         self
     }
